mute/ban debugging
handling scenario where user has mute role or is banned, but no infraction is in the db. we don't want the bot getting confused here, but we also don't want to dispatch two events to the database too quickly. so the resolution is to ask what the moderator would like to do.